Provide chrome headless driver
This adds in headless chrome as the default driver for acceptance tests.
The driver is added in rails_helper, along with the prior full chrome
option as well. This will allow a developer to change the driver to
:chrome from :headless_chrome if they want to see the test execute for
debugging purposes.

Additional next steps are provided at the end to document the additional
system dependencies that this requires to run locally.