Use FactoryGirl parent strategy
This change modifies FactoryGirl's configuration to inherit the strategy
to construct objects from the parent object for associations. Prior to
this change, and the FactoryGirl default, is to `create` all associated
records from the factory, regardless of the strategy of the parent
object.

This previous default resulted in records being persisted to the
database when one wouldn't expect them to be.