Prepare application to use system tests
This change removes the need for the DatabaseCleaner dependency, and
includes default configuration for utilizing rails' system tests. These
should be used for acceptance tests rather than rspec feature tests.

These tests are now built into rails, so it includes a nice DSL for
registering drivers, and also will operate in the same process as other
tests, so there's no need for a tool like DatabaseCleaner, and
transactional fixtures may be used for acceptance tests as well as other
test types.