Limit logging in test environment
This sets the rails logger in test to point to nothing, and raises the
logging level to fatal so that only the highest severity even attempts
to log.

There is an ENV VAR that can be set to override this, if the default
logger should be used in a particular instance. This should decrease
the run time of the test suite for the resulting applications.