Drop pg gem constraints
Starting in Rails 5.1.5, the pg gem should now work with version 1.0. As
a result, this template no longer needs to lock pg to pre-1.0. In fact,
running this template file as-is against Rails 5.1.5 will result in an
error, as we're trying to specify conditions on pg twice: once by rails'
default, and again with this now-deleted line in this template.