rspec-rails Generator
This updates gnarails to depend on rspec-rails, and changes the
generator install to use the available `generate` method in the rails
template. Adding rspec-rails to the gemspec ensures that it is available
within the gem and on the system running the template.

Additionally, this modifies the rails development environment so that
the listen gem is disabled prior to running the generator, and enabled
after the generator has run. When executing in a subshell, listen being
turned on was causing issues. To get around this, it's temporarily
disabled while running the template, but turned back on so it's not
modified to the end user when the rails app is ready for them to use.