Run integration spec in CI
This updates the circle configuration so that the integration spec from
the sample application may be run, differently from the local setup.

This involves creating a workflow, which is multiple CI jobs put
together, to achieve the same result.
* The first job will run the unit specs for the gnarails project.
* The second job pulls a non-circle ruby image to generate a rails app
using gnarails. This is necessary because it appears the circle ruby
image includes some work that vendors gems in some way that running
`gnarails new` doesn't allow you to access the gems from the rails
session, just from gnarails itself. The result of that rails app is
stored in a workspace, which allows you to share files across jobs in a
workflow.
* The final job grabs the rails app from the shared workspace, runs the
specs, and other configuration items (coverage analysis, security
analysis, etc) that is in the Circle config you get from gnarails to
ensure that creating a new rails app from gnarails will work as
expected.