Test N+1 query protection
This introduces a new model into the test application - a comment. A job
posting has many comments. This is used to intentionally introduce a N+1
query on the job posting index page (displaying all the comments without
preloading the comments).

An additional system test is provided which will error out, due to an
N+1 query. The bullet gem should be configured such that if any code
path is executed in a test, the test will raise an error and fail. This
test passes because the test itself is ensuring that the appropriate
error is raised when an N+1 is encountered.