Update to ruby 2.5.0
This specifies ruby 2.5.0 as the latest ruby version to use with the
created rails application. This also extracts all templated uses of the
ruby version into a constant, so for the template, only one location is
needed to be updated to change the ruby version.

There are still other places, like the circle config of the project
itself, the .ruby-version file, and the gemspec where the ruby version
will need to be changed along with the constant.