Support rails generator options
This explicitly sets all of the different options from creating a new
rails app that a gnarails application will accept. This includes all
options that are available when generating a rails application, while
excluding options that aren't applicable when using this template, such
as the database option or the template option. These are explicitly set
by using gnarails, and as such, are not available for users of this
generator.