Update README template instructions
This changes the README instructions for using the template back to
pulling down the project locally and using the template file locally
rather than pulling from the github HTTP location.

The reason for this is that the template relies on particular files
being available at certain paths/locations, and pulling the template
from the HTTP resource does not allow for that.