Use application's Gemfile, not gnarrails Gemfile
Currently when generating an app, Gnarails Gemfile is used which, with
the exception of rspec-rails has all of the dependencies necessary to
run the Rails app which makes this bug harder to find.

When Bundler starts it looks up the directory tree for a `Gemfile` and
finds `gnarails` Gemfile first. This isn't a problem except that it's
cached in the `BUNDLE_GEMFILE` environment variable. This means each
command we run on the generated app is trying to use gnarails `Gemfile`
instead of its own. Because it's an environment variable new shells
(like running commands in a Ruby script via backticks) inherit this
value and further hide the root cause of the strange behavior.

To resolve this issue, once we start running commands on the generated
app we reset the `BUNDLE_GEMFILE` environment variable to clear the
cache and allow Bundler to find the generated applications Gemfile.