Fix dependency issues and CI
This removes cache from CI since gems are vendored in folders, things
break due to having to work with multiple Gemfiles and project roots.

This also moves most of the actual behavior of gnarails to happen after
bundle installing instead of before which is seemingly related to the
gem overshadowing issue.