Use Pronto's GH Review formatter
This modifies the formatter that pronto uses to comment on syle guide
violations on a PR. The Review formatter should consolidate all feedback
in a single Github review, rather than individual comments for each
violation.

This should cut down on email traffic and notifications, as all feedback
from the source will designate from a single review.

This change affects both pronto's usage in this repo, as well as the
setup for any rails app generated using gnarails.