Remove Usage of --rdoc and --ri
Rubygems 3.0.0 removed the --rdoc and --ri options from install and
update, which were deprecated previously. This was removed in the
following pull request: https://github.com/rubygems/rubygems/pull/2354.

This removes the use of these options in favor of the new `--no-document`
option. The documentation states that this option will,
"disable documentation generation".