Upgrade ruby to 2.6.1
This updates the version of ruby tested in the application to be ruby
2.6.1. This still relies on the `RUBY_VERSION` constant that's defined
in ruby to know what version of ruby to include.

This also removes the code that will add the ruby version to the
Gemfile. Rails 5.2 will add the ruby version to the Gemfile for you,
based on this contribution: https://github.com/rails/rails/pull/30016.

The latest Circle images utilize bundler >= 2, which is causing issues
for the template install process. This circumvents that by modifying the
image to remove the currently-installed version of bundler, and
replacing it with a specific version that's known to work. Additional
effort will be required to support bundler >= 2.