Upgrade ruby to 2.6.0
This updates the version of ruby tested in the application to be ruby
2.6.0. This still relies on the `RUBY_VERSION` constant that's defined
in ruby to know what version of ruby to include.

This also removes the code that will add the ruby version to the
Gemfile. Rails 5.2 will add the ruby version to the Gemfile for you,
based on this contribution: https://github.com/rails/rails/pull/30016.