Raise error on missing translations in dev and test. (#144)
This PR adds config.action_view.raise_on_missing_translations = true to the default config files for dev and test. This will throw errors for missing translations instead of using fallbacks by default.

This PR also removes the chromedriver-helper gem from the generated Gemfile, as this causes errors when running on CI. Regardless, that gem is deprecated, and we instruct the user to install chromedriver in the post-install message anyways.