Handle change of CircleCI images after ruby 2.7.0
**This Commit**
Updates the template's `.circleci/config.yml` to conditionally determine
the CircleCI image name for ruby.

**Why?**
CircleCI switched its images from `stretch-...` to `buster-...` with the
[new release of
Debian](https://discuss.circleci.com/t/docker-convenience-images-updates-july-2019/31239).
Because of this, if the user generated a new rails app with ruby 2.7.x
installed, they would attemp to reference the image `2.7.x-stretch-node-browsers`
which isn't available. Instead, it should be
[`2.7.x-buster-node-browsers`](https://hub.docker.com/r/circleci/ruby/tags?page=1&name=2.7).