Update postgres version in CircleCI config
**This Commit**
Updates the postgres version in the CircleCI config to be 13.0

**Why?**
Because new users with new DBs will probably not be running 9.6.5 so
having CI use it is unlikely to match their environment. They may use
new features that break in CI because of the old version. Also, who
knows, maybe the new version is faster and everyone's CI will be better.

**Note**
There are two methods for handling auth for these postgres images
described
[here](https://discuss.circleci.com/t/postgresql-image-password-not-specified-issue/34555).
I chose to just disable the auth as it matches what we had before but we
could also set up some username/password in the database.yml and match
that (I imagine).