[skip ci] Passthrough args to rails new directly
**This Commit**
Does a super-hack job of sort of laying out a method for passing through
all command line args to `rails new`.

**Why?**
1. `gnarails` doesn't accept all the options `rails new` does (I think
the one I ran into was `webpack` or `webpacker` or something, it was
forever ago).
2. `gnarails` probably doesn't want to reinvent `rails new` arguments
for the rest of time

**Note**
This is obviously a super ugly hack. This is basically ejecting from
`thor`. Which is maybe fine since we're basically just sliding a couple
of args into the grand scheme of things. Although, I imagine, the
current setup:
1. doesn't work (I don't know how to test it locally)
2. doesn't allow the user to unset things (I don't know what `rails new`
does with `--database=postgresql --database=something_else` but maybe it
does the right thing)