Remove extraneous entries from Gemfile
When generating a new Rails app, I was confronted with an error that capybara
(and later selenum-webdriver) could not be specified twice with different
version requirements in a Gemfile. The one gnarails was adding had no
specification and the one Rails specified a version. So, since they're being
added as part of `rails new`, we don't need them here.